requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
aiohttp==3.8.6
asyncio
python-dotenv==1.0.0
